Started Level Manager
Created a Level Manager class which will be responsible for the creation and maintenance of the 3 main parts of every level: The monster spawner, the play area, and the end zone.

Currently, the Level manager only has functionality to create and populate the play area with the base block type.

Next steps are to implement an improved system that will fill out the flooring and the buildable layers separately (using floor blocks and base blocks respectively).