Level Manager Bugfix
Fixed a bug with the level manager where the 2D layer arrays were not tracking blocks correctly if the play area was not a square.

Blocks would be generated and placed as expected, but the arrays were not being initialised to the correct sizes. As such, blocks lying outside of the "square" zone could not be accessed for deletion when resizing.